Adding replacing PR files logic

+++ bot/util_clone_wrapper.py
@@ -13,12 +13,24 @@ import os.path
 
 from coverage_calc.coverage_helper import coverage_helper
 from util_duplicates_checker import util_duplicates_checker
+from utils_crabot import get_list_changed_files
 
 PR_COMMENT="issue_comment"
 COMMIT_COMMENT="commit_comment"
 COMMENT_ON_PR="pull"
 PR="pull_request"
 TMP_DIR_NAME = 'tmp_clone'
+
+def extract_result(data):
+	result = {}
+	for item in data:
+		relative_path = item['filename']
+		if('.java' in relative_path):
+			#strip relative path of filename
+			relative_path = relative_path.rsplit('/', 1)[0]
+			raw_url = item['raw_url']
+			result[relative_path] = raw_url
+	return result
 
 
 '''
@@ -32,45 +44,65 @@ description
 	Returns text with overall results
 '''
 def util_clone_wrapper(dict_payload, request_type, coverageFlag, duplicateFlag):
-	if(request_type == COMMENT_ON_PR or request_type==COMMIT_COMMENT or request_type==PR):
-		pr_url = str(dict_payload['issue']['pull_request']['url'])
+	list_files=get_list_changed_files(dict_payload,request_type)
+	result=extract_result(list_files)
 
-		pull_json_text =  requests.get(pr_url)
-		data = json.loads(pull_json_text.content)
+	if result:
+		if(request_type == COMMENT_ON_PR or request_type==COMMIT_COMMENT or request_type==PR):
+			pr_url = str(dict_payload['issue']['pull_request']['url'])
 
-		repouri = data['head']['repo']['clone_url']
-		reponame = data['head']['repo']['name']
+			pull_json_text =  requests.get(pr_url)
+			data = json.loads(pull_json_text.content)
 
-		try:
-			call('mkdir '+TMP_DIR_NAME, shell=True)							# make temp dir
-			call("git clone " + repouri, shell=True, cwd='./'+TMP_DIR_NAME) 	# clone repo in temp
+			repouri = data['head']['repo']['clone_url']
+			reponame = data['head']['repo']['name']
 
-			cmdlist = list()
-			cmdlist.append("git checkout mvn")				#TODO remove later
-			for cmd in cmdlist:
-				call(cmd, shell=True, cwd='./' + TMP_DIR_NAME + '/'+reponame) 		# run util on 
+			try:
+				if os.path.exists(TMP_DIR_NAME):
+					call('rm -rf ' + TMP_DIR_NAME, shell=True)	
 
-			############################
-			# NOW WE HAVE CODE CLONED INTO TMP_DIR_NAME FOLDER
-			# WE CAN RUN REQUESTED ANALYSIS NOW
-			# CALL MODULES, and GET RESULTS
-			############################
-			results = ""
-			if (coverageFlag):
-				results += coverage_helper(TMP_DIR_NAME, reponame)
+				call('mkdir '+TMP_DIR_NAME, shell=True)							# make temp dir
+				call("git clone " + repouri, shell=True, cwd='./'+TMP_DIR_NAME) 	# clone repo in temp
 
-			if (duplicateFlag):
-				results += util_duplicates_checker('./' + TMP_DIR_NAME + '/'+reponame)
+				REPO_PATH = './' + TMP_DIR_NAME + '/'+reponame
 
-			return results
+				cmdlist = list()
+				cmdlist.append("git checkout mvn")				#TODO remove later
+				for cmd in cmdlist:
+					call(cmd, shell=True, cwd=REPO_PATH) 		# run util on 
+
+				#replace files from result dictionary
+				for relative_path, raw_url in result.iteritems():
+					call('cd ' + REPO_PATH + '/' + relative_path, shell=True) 
+					call("curl -H 'Accept: application/vnd.github.v3.raw' -O -L '" + raw_url, shell = True)		#downlaod & replace files
+					call('cd ' + REPO_PATH, shell=True)		#back to src directory of repo
+
+				############################
+				# NOW WE HAVE CODE CLONED INTO TMP_DIR_NAME FOLDER
+				# WE CAN RUN REQUESTED ANALYSIS NOW
+				# CALL MODULES, and GET RESULTS
+				############################
+				results = ""
+				if (coverageFlag):
+					results += str(coverage_helper(TMP_DIR_NAME, reponame))
+
+				if (duplicateFlag):
+					results += str(util_duplicates_checker(REPO_PATH))
+
+				if results and not results.isspace():			
+					return results
+				else:
+					return "No results found for coverage and duplicates checks!"
 
 
 
-		except (RuntimeError, TypeError, NameError) as ex:
-			print 'exception occurred'
-			print ex
+			except (RuntimeError, TypeError, NameError) as ex:
+				print 'exception occurred'
+				print ex
 
-		finally:
-			print 'deleting temp dir'
-			call('rm -rf ' + TMP_DIR_NAME, shell=True)							# remove temp	
+			finally:
+				print 'deleting temp dir'
+				call('rm -rf ' + TMP_DIR_NAME, shell=True)							# remove temp	
+	else:
+		return "No java file found in Pull Request!\n"
